// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Plan = {
  FREE: 'FREE',
  BASIC: 'BASIC',
  PRO: 'PRO'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]